+++ src/App.jsx
@@ -1,22 +1,13 @@
-// todo **`` Create a new branch each week!!
-
+import AddTodoForm from "./AddTodoForm";
 import "./App.css";
-
-const todoList = [
-  { id: 1, title: "Go to store" },
-  { id: 2, title: "Read book" },
-  { id: 3, title: "Buy a monkey" },
-];
+import TodoList from "./TodoList";
 
 function App() {
   return (
     <>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map(({ id, title }) => (
-          <li key={id}>{title}</li>
-        ))}
-      </ul>
+      <AddTodoForm />
+      <TodoList />
     </>
   );
 }
